requirements.txt
asgiref==3.7.2
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Django==4.2.6
djangorestframework==3.14.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.17.0
MarkupSafe==2.1.3
msgpack==1.0.7
numpy==1.26.1
pandas==2.1.1
psutil==5.9.6
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
